[Presentation] QuotesPresenterTest
