--- presentation/src/test/kotlin/io/github/dector/quotes/AssertUtils.kt
@@ -1,5 +1,0 @@
-package io.github.dector.quotes
-
-inline infix fun Any?.equalTo(other: Any?): Boolean = this?.equals(other) ?: other == null
-
-inline fun Any?.nonNull(): Boolean = this != null
